+++ src/cli/send/scrt/Scrt.definition.ts
@@ -1,0 +1,40 @@
+/**
+ * This program and the accompanying materials are made available and may be used, at your option, under either:
+ * * Eclipse Public License v2.0, available at https://www.eclipse.org/legal/epl-v20.html, OR
+ * * Apache License, version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
+ *
+ * Copyright Contributors to the Zowe Project.
+ */
+
+import { ICommandDefinition } from "@zowe/imperative";
+import { ProfileConstants } from "../../../api/ProfileConstants";
+
+/**
+ * This definition is of imperative type "command" and therefore must have a
+ * command handler (which performs the "work" for this command).
+ *
+ * Property Summary:
+ * =================
+ * "name" of the [object]. Should be a noun (e.g. data-set)
+ * "aliases" normally contains a shortened form of the command
+ * "summary" will display when issuing the help on this [objects] [action]
+ * "type" is "command" which means a handler is required
+ * "handler" is the file path to the handler (does the work)
+ * "options" an array of options
+ */
+export const ScrtDefinition: ICommandDefinition = {
+    name: "scrt",
+    aliases: [],
+    summary: "Test the transmission of SCRT data from client to server",
+    description: "Sends a GET REST request with an SCRT header to the targeted REST service. " +
+        "The purpose of this command is to test the transmission of SCRT data to a REST service. " +
+        "Note that an SCRT header is not sent when using a direct connection to z/OSMF.",
+    type: "command",
+    handler: __dirname + "/Scrt.handler",
+    profile: {
+        optional: ["scrtProfile"],
+    },
+    options: [...ProfileConstants.SAMPLE_SCRT_OPTIONS]
+};