+++ src/cli/send/scrt/ScrtRestClient.ts
@@ -1,0 +1,130 @@
+/**
+ * This program and the accompanying materials are made available and may be used, at your option, under either:
+ * * Eclipse Public License v2.0, available at https://www.eclipse.org/legal/epl-v20.html, OR
+ * * Apache License, version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
+ *
+ * Copyright Contributors to the Zowe Project.
+ */
+
+import {
+    IImperativeError, Logger, RestClient,
+    RestConstants, SessConstants,
+} from "@zowe/imperative";
+import { ZosmfHeaders } from "@zowe/core-for-zowe-sdk";
+
+/**
+ * Wrapper for the RestClient to perform common error handling.
+ * @export
+ * @class ScrtRestClient
+ * @extends {RestClient}
+ */
+export class ScrtRestClient extends RestClient {
+
+    /**
+     * Use the Zowe logger instead of the imperative logger
+     * @type {Logger}
+     */
+    public get log(): Logger {
+        return Logger.getAppLogger();
+    }
+
+    /**
+     * Append z/OSMF specific headers to the caller's headers.
+     * This is done so that a z/OSMF url can be used for testing the
+     * transmission of SCRT headers. The X_CSRF_ZOSMF_HEADER will be
+     * harmlessly ignored by REST services other than z/OSMF.
+     * @param {any[] | undefined} headers - current header array
+     * @memberof ZosmfRestClient
+     */
+    protected appendHeaders(headers: any[] | undefined): any[] {
+        if (headers == null) {
+            headers = [ZosmfHeaders.X_CSRF_ZOSMF_HEADER];
+        } else {
+            headers.push(ZosmfHeaders.X_CSRF_ZOSMF_HEADER);
+        }
+        return headers;
+    }
+
+    /**
+     * Process an error encountered in the rest client
+     * @param {IImperativeError} original - the original error automatically built by the abstract rest client
+     * @returns {IImperativeError} - the processed error with details added
+     * @memberof ScrtRestClient
+     */
+    protected processError(original: IImperativeError): IImperativeError {
+        let causeErrorsJson;
+        let causeErrorsString = "";
+        if (original.causeErrors) {
+            causeErrorsString = original.causeErrors;
+        }
+        try {
+            // don't try to parse an empty string
+            if (causeErrorsString !== "") {
+                causeErrorsJson = JSON.parse(causeErrorsString);
+                // if we didn't get an error trying to parse causeErrorsString, check if there is a stack
+                // on the JSON error and delete it
+                if (causeErrorsJson.stack != null) {
+                    this.log.error("An error was encountered which contains a stack trace." +
+                        " Here is the full error before deleting the stack:\n%s", JSON.stringify(causeErrorsJson));
+                    this.log.error("The stack has been deleted from the error before displaying the error to the user");
+                    delete causeErrorsJson.stack; // remove the stack field
+                    original.causeErrors = JSON.stringify(causeErrorsJson, null);
+                }
+            }
+        } catch (err) {
+            // if there's an error, the causeErrors text is not JSON
+            this.log.debug(
+                "Encountered an error trying to parse causeErrors as JSON  - causeErrors is likely not JSON format\n" +
+                "Reason: " + err.message
+            );
+        }
+
+        const origMsgFor401 = original.msg;
+        // extract properties from causeErrors and place them into 'msg' as user-focused messages
+        if (causeErrorsJson?.details?.length > 0) {
+            for (const detail of causeErrorsJson.details) {
+                original.msg += "\n" + detail;
+            }
+        }
+        if (causeErrorsJson?.messages?.length > 0) {
+            for (const message of causeErrorsJson.messages) {
+                original.msg += "\n" + message.messageContent;
+            }
+        }
+        if (causeErrorsJson?.message?.length > 0) {
+            original.msg += "\n" + causeErrorsJson.message;
+        }
+
+        // add further clarification on authentication errors
+        if (this.response && this.response.statusCode === RestConstants.HTTP_STATUS_401) {
+            if (!original.causeErrors || Object.keys(original.causeErrors ).length === 0) {
+                /* We have no causeErrors, so place the original msg we got for a 401
+                 * into the 'response from service' part of our error.
+                 */
+                original.causeErrors = `{"Error": "${origMsgFor401}"}`;
+            }
+            original.msg  += "\nThis operation requires authentication.";
+
+            if (this.session.ISession.type === SessConstants.AUTH_TYPE_BASIC) {
+                original.msg += "\nUsername or password are not valid or expired.";
+            } else if (this.session.ISession.type === SessConstants.AUTH_TYPE_TOKEN) {
+                if (this.session.ISession.tokenType === SessConstants.TOKEN_TYPE_APIML && !this.session.ISession.basePath) {
+                    original.msg += `\nToken type "${SessConstants.TOKEN_TYPE_APIML}" requires base path to be defined.\n` +
+                        "You must either connect with username and password or provide a base path.";
+                } else {
+                    original.msg += "\nToken type = '" + this.session.ISession.tokenType +
+                        "' is not valid, token is invalid, or token is expired.\n" +
+                        "To obtain a new valid token, use the following command: `zowe config secure`\n" +
+                        "For CLI usage, see `zowe config secure --help`";
+                }
+            // TODO: Add PFX support in the future
+            } else if (this.session.ISession.type === SessConstants.AUTH_TYPE_CERT_PEM) {
+                original.msg += "\nCertificate is not valid or expired.";
+            }
+        }
+
+        return original;
+    }
+}